tests: Fix convenience function to run gpg.
* tests/openpgp/defs.scm (pipe:gpg): Remove stray dash.
--

The extra dash is problematic at export and import because this does
not export a single key as intended by the tests but does two key
lookups where the second fails because that userid is always "-".  For
the current keyring and keybox code this does not matter because it is
ignored.  The keyboxd code however is more picky and returns an error.

Fixes-commit: 9609cb20e4caee739b9fc4fd36797029d2970041
Signed-off-by: Werner Koch <wk@gnupg.org>